Refactor and streamline repository structure and functionality
Significant reorganization undertaken: introduced better modular coding practices with new function definitions and splitting into smaller subfiles. Removed legacy code (e.g., process.py), replaced handlers for improved performance, and updated imports appropriately. Expanded the `.gitignore` and added relevant badges.
