+++ pyproject.toml
@@ -35,6 +35,7 @@ dependencies = [
     "scipy==1.15.2",
     "fiona==1.10.1",
     "omnicloudmask==1.0.11",
+    "geopandas==1.0.1",
 ]
 
 [project.optional-dependencies]
