+++ spectralmatch/__init__.py
@@ -1,6 +1,16 @@
-from spectralmatch.process import global_match, local_match
+from .match.global_regression import global_regression
+from .match.local_block_adjustment import local_block_adjustment
+from .handlers import merge_rasters, align_rasters
+from .mask import create_cloud_mask_with_omnicloudmask, post_process_raster_cloud_mask_to_vector, create_ndvi_mask, post_process_threshold_to_vector
 
 __all__ = [
-    "global_match",
-    "local_match",
+    "global_regression",
+    "local_block_adjustment",
+    "merge_rasters",
+    "create_cloud_mask_with_omnicloudmask",
+    "post_process_raster_cloud_mask_to_vector",
+    "create_ndvi_mask",
+    "post_process_threshold_to_vector",
+    "merge_rasters",
+    "align_rasters"
 ]
